feat: log attachment names in delete hook
If a message has attachments when deleted (such as an image, a GIF, a
video or a general file), the names and content types will be logged as
well.

The intention behind this is to stop the "Partial message" errors
whenever a message has no content but still attachments. Since it is
legal to avoid typing content into a message as long as it has a file,
it is not valid to just check for the emptyness of the message content
to treat it as partial. Empty bodies with attachments are not partial
messages.

Also, refactor the file to make it strict.