chore: update DiscordJS to 13-LTS and NodeJS to 20
This package is finally running of the 13-lts version, which means that
this is the latest version of the 13 series. Could be upgraded to 14, or
now that this version has lost so many unused features, could be
rewritten in a better programming language.

NodeJS has been updated to 20, since 13-lts requires a newer version of
NodeJS anyway. The reason on choosing 20 and not 22 is because some
dependencies are asking about features that have been removed from
NodeJS 22 (punycode).