refactor: use a separate task for karma levels
Rather than asserting that the level is correct after every message sent
to a channel, this commit will install a separate loop that checks for
that.

So now when you send a message to a karma enabled channel, all that is
going to happen is that the member and the last channel of user activity
will be added into a queue. The last channel is added in order to
possibly submit a "level up" message. However, only one channel per
pending member is acceptable, if multiple messages are sent, only the
most recent channel will be taken into consideration when deciding where
to send the "level up" message.

This queue is consumed in a separate event loop every couple of seconds.
The reason behind this change is to prevent the new message hook to do a
lot of complex SQLite queries on every message, and instead just use the
message hook to add a row into the table. All the sums are done in a
separate thread, and only once per user every couple of seconds,
preventing a SQL query storm when an user sends multiple messages in a
short span of time.

Server telemetry and metrics will indicate whether this keeps the
database and server CPU more calm.